config: increase MAX_MESSAGE_LENGTH
MAX_MESSAGE_LENGTH imposes limit on file size. Default value of 1024 is
not enough to store some RSA2048 certificates. Limit is now 2048 bytes.
Note that we cannot increase limit forever, due to how Trussed APIs work
we always need to allocate buffer of MAX_MESSAGE_LENGTH size even we are
going to use only part of it. If we increased it too much we would run
out of memory very quickly.